+++ R/getFunctions.R
@@ -125,6 +125,28 @@ getCRPS <- function(z.pred, mean.pred, sd.pred){
   
 }
 
+#' Based on a specific taper scale (delta), retrieves the density of the covariance matrix.
+#' @description Based on a specific taper scale (delta), retrieves the density of the covariance matrix.
+#'
+#' @usage getDensityFromDelta(coco.object, delta)
+#' @param coco.object \code{(S4)} a fitted [coco()] object.
+#' @param delta \code{(numeric)} a delta taper scale (delta).
+#' @returns (\code{numeric vector}) the associate density of the tapered covariance matrix.
+#' @author Federico Blasi
+getDensityFromDelta <- function(coco.object, delta){
+  
+  if(coco.object@type != "sparse"){stop("only for sparse coco objects.")}
+  
+  if(length(coco.object@output) == 0){
+    coco.object@output$par <- rep(0, .cocons.get.npars(coco.object))
+  }
+  
+  coco.object@info$delta <- delta
+  
+  return(summary(getCovMatrix(coco.object))$density)
+  
+}
+
 #' Evaluates the spatially-varying functions from a coco object at locs
 #' @description Evaluates the spatially-varying functions of the nonstationary spatial structure.
 #'
@@ -188,14 +210,14 @@ getSpatEffects <- function(coco.object){
 
 }
 
-#' Computes the spatial trend of a (fitted) coco object
-#' @description Compute the trend of the (fitted) coco object.
+#' Computes the spatial mean of a (fitted) coco object
+#' @description Computes the spatial mean of the (fitted) coco object.
 #'
-#' @usage getTrend(coco.object)
+#' @usage getSpatMean(coco.object)
 #' @param coco.object \code{(S4)} a fitted coco S4 object.
 #' @returns (\code{numeric vector}) a vector with the adjusted trend.
 #' @author Federico Blasi
-getTrend <- function(coco.object){
+getSpatMean <- function(coco.object){
   tmp_scaled <- getScale(coco.object)$std.covs
   return(tmp_scaled %*% getEstims(coco.object)$mean)
 }
@@ -203,28 +225,30 @@ getTrend <- function(coco.object){
 #' Compute approximate confidence intervals for a coco object
 #' @description Compute approximate confidence intervals for a (fitted) coco object.
 #'
-#' @usage getCIs(coco.object, inv.hess, alpha = 0.05)
+#' @usage getCIs(coco.object, inv.hess, alpha = 0.95)
 #' @param coco.object \code{(S4)} a fitted coco S4 object.
 #' @param inv.hess \code{(matrix)} Inverse of the Hessian. \link{getHessian}.
 #' @param alpha \code{(numeric)} confidence level.
 #' @returns (\code{numeric matrix}) a matrix with approximate confidence intervals for each parameter in the model.
 #' @author Federico Blasi
-getCIs <- function(coco.object, inv.hess, alpha = 0.05){
+getCIs <- function(coco.object, inv.hess, alpha = 0.95){
   
   if(alpha >= 1 || alpha <= 0){stop("check alpha.")}
   
-  tmp_par.pos <- cocons::getDesignMatrix(coco.object@model.list,coco.object@data)$par.pos
+  tmp_DM <- cocons::getDesignMatrix(coco.object@model.list,coco.object@data)
+  
+  tmp_par.pos <- tmp_DM$par.pos
   
   Hess_modified <- cocons::getModHess(coco.object = coco.object, 
                                      inv.hess = inv.hess)
   
   estims <- unlist(getEstims(coco.object)[which(unlist(lapply(tmp_par.pos,is.logical)))])[unlist(tmp_par.pos[which(unlist(lapply(tmp_par.pos, is.logical)))])]
   
-  to_return <- matrix(c(estims, estims), ncol = 2, byrow = FALSE) + as.matrix(sqrt(diag(Hess_modified)), ncol = 1) %*% matrix(c(-1, 1) * stats::qnorm(1 - alpha/2) ,ncol = 2)
+  to_return <- matrix(c(estims, estims), ncol = 2, byrow = FALSE) + as.matrix(sqrt(diag(Hess_modified)), ncol = 1) %*% matrix(c(-1, 1) * stats::qnorm(alpha) ,ncol = 2)
   
-  rownames(to_return) <- names(estims)
-  
-  colnames(to_return) <- c(paste((alpha/2 )* 100, "%"),paste((1 - alpha/2 ) * 100, "%"))
+  rownames(to_return) <- .cocons.updateNames(names(estims),tmp_DM$model.matrix)
+
+  colnames(to_return) <- c(paste(( (1 - alpha)/2 )* 100, "%"),paste( (1 - ( 1 - alpha )/2) * 100, "%"))
   
   return(to_return)
 }
@@ -821,7 +845,7 @@ getBoundariesV3 <- function(coco.object,
 }
 
 #' getHessian
-#' @description numerically approximate the Hessian. Hessians of parameters based on "pmle" are based on full likelihoods.
+#' @description numerically approximate the Hessian. Hessians of parameters based on "pml" are based on full likelihoods.
 #' @usage getHessian(coco.object, ncores = parallel::detectCores() - 1, 
 #' eps = .Machine$double.eps^(1/4))
 #' @param coco.object \code{(S4)} a fitted coco object.
@@ -833,6 +857,8 @@ getHessian <- function(coco.object, ncores = parallel::detectCores() - 1,
                        eps = .Machine$double.eps^(1/4)){
   
   if(coco.object@type == "dense"){
+    
+    if(coco.object@info$optim.type == "reml"){stop("reml hessian not implemented yet.")}
     
     p <- base::length(coco.object@output$par)
     H <- base::matrix(NA, p, p)
@@ -860,7 +886,7 @@ getHessian <- function(coco.object, ncores = parallel::detectCores() - 1,
     
     x_covariates <- coco_x_std$std.covs
     
-    if(coco.object@info$optim.type == "pmle"){
+    if(coco.object@info$optim.type == "pml"){
       
       f00 <- cocons::GetNeg2loglikelihood(coco.object@output$par, par.pos = par.pos,
                                    locs = coco.object@locs,
@@ -878,7 +904,7 @@ getHessian <- function(coco.object, ncores = parallel::detectCores() - 1,
     parallel::clusterEvalQ(cl, library("cocons"))
     
     parallel::clusterExport(cl = cl, list("x_covariates", "coco.object",
-                                          "n", "par.pos", "eps","f00"),
+                                          "n", "par.pos", "eps", "f00"),
                             envir = environment())
     
     vector_responses <- parallel::parApply(cl = cl, 
@@ -972,7 +998,7 @@ getHessian <- function(coco.object, ncores = parallel::detectCores() - 1,
     
     cholS <- spam::chol.spam(ref_taper)
     
-    if(coco.object@info$optim.type == "pmle"){
+    if(coco.object@info$optim.type == "pml"){
       
       f00 <- cocons::GetNeg2loglikelihoodTaper(theta = coco.object@output$par,
                                                par.pos = par.pos,