+++ NEWS.md
@@ -1,3 +1,33 @@
+## cocons 0.1.4
+
+- Documentation polished
+
+### Enhances
+
+- `GetNeg2loglikelihood`, `GetNeg2loglikelihoodTaper`, and `GetNeg2loglikelihoodTaperProfile` now much faster (~35\% faster)
+- `cocoOptim`:
+  - new `REML` estimation for dense `coco` type.
+- now cpp functions `cov_rns` and `cov_rns_taper` include as special cases the well-known shapes when nu = 0.5, 1.5, and 2.5, yielding computational speed-ups.
+- now `getCIs` display names of covariate instead of index of Design Matrix.
+- now `.cocons.check.info` also checks for smoothness model and smooth.limits.
+  
+### Changes
+
+- `coco` now accepts `data.frame` for locs argument, which is then converted to matrix.
+- `getTrend` now is called `getSpatMean`.
+- `cocoPredict` renamed output spatial mean vectors: `trend` is now called `systematic`, while `mean` is called `stochastic`.
+- `cocoOptim` : 
+  - reordering of arguments.
+  - `mle` estimation method now called `ml`.
+  - `pmle` estimation method now called `pml`.
+- `getHessian` :
+  - `mle` to `ml` and `pmle` to `pml`.
+- `getCIs` now `alpha` argument reflects confidence level instead of 1-confidence level.
+
+### Fixes
+
+- fix a bug for plotOptimInfo when handling `pml` or `reml` objects.
+
 ## cocons 0.1.3
 
 - Vignette updates, watermark removed